+++ src/Label.js
@@ -79,7 +79,8 @@ const Label = (props) => {
     check && inline && disabled ? 'disabled' : false,
     size ? `col-form-label-${size}` : false,
     colClasses,
-    colClasses.length ? 'col-form-label' : false
+    colClasses.length ? 'col-form-label' : false,
+    !check && !colClasses.length ? 'form-control-label' : false
   ), cssModule);
 
   return (