+++ src/__tests__/Nav.spec.js
@@ -21,6 +21,30 @@ describe('Nav', () => {
     expect(wrapper.html()).toBe('<ul class="nav">Children</ul>');
   });
 
+  it('should handle justified attribute', () => {
+    const wrapper = shallow(<Nav justified />);
+
+    expect(wrapper.html()).toBe('<ul class="nav nav-justified"></ul>');
+  });
+
+  it('should handle pills attribute', () => {
+    const wrapper = shallow(<Nav pills />);
+
+    expect(wrapper.html()).toBe('<ul class="nav nav-pills"></ul>');
+  });
+
+  it('should handle tabs attribute', () => {
+    const wrapper = shallow(<Nav tabs />);
+
+    expect(wrapper.html()).toBe('<ul class="nav nav-tabs"></ul>');
+  });
+
+  it('should handle vertical attribute', () => {
+    const wrapper = shallow(<Nav vertical />);
+
+    expect(wrapper.html()).toBe('<ul class="nav flex-column"></ul>');
+  });
+
   it('should pass additional classNames', () => {
     const wrapper = shallow(<Nav className="extra" />);